requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
Pillow==10.1.0
pymongo==4.6.0